debug
solve the problem of "blocks that is larger than 128 cannot merge" problem.

+++ src/edu/scu/Game.java
@@ -105,7 +105,7 @@ public class Game {
 				list.remove(0);
 				continue;
 			}
-			else if(list.get(0)==list.get(1)) {
+			else if((int)list.get(0)==(int)list.get(1)) {
 				//merge 2 numbers,then pop out two numbers in the list
 				res[pos++]=list.get(0)*2;
 				list.remove(1);
